Разделить написанный код на модули
