+++ js/string-length.js
@@ -1,0 +1,3 @@
+//Функция, проверяющая длину строки относительно максимально возможной
+//const checkLength = (checkedString, MAX_LENGTH) => checkedString.length <= MAX_LENGTH;
+//checkLength();
